requirements.txt
mkdocs>=1.6
mkdocs-material>=9.5
pymdown-extensions>=10.7
PyYAML>=6.0
